plugins/python:  * Fixed Linux compilation issue introduced in previous commit
git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@3557 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ trunk/openvibe-plugins/python/trunc/src/ovpCBoxAlgorithmPython.cpp
@@ -477,7 +477,7 @@ OpenViBE::boolean CBoxAlgorithmPython::initialize(void)
 		this->getLogManager() << LogLevel_Error << "Failed to run [" << m_sScriptFilename << "]";
 		if(l_pResult) 
 		{
-			this->getLogManager() << ", result = " << PyInt_AsLong(l_pResult) << "\n";
+			this->getLogManager() << ", result = " << (int32)PyInt_AsLong(l_pResult) << "\n";
 		} 
 		else
 		{
